rename ConnectedComponent files
