+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DirectMessageSenderComponent } from './direct-message-sender/direct-message-sender.component';
-import { ConnectedComponent } from './ConnectedComponent/connected.component';
+import { ConnectedComponent } from './connected-component/connected.component';
 import { LoginComponent } from './login-component/login.component';
 
 @NgModule({
